# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A lightweight backend RESTful API designed for tracking and managing personal or organizational expenses. This API provides secure endpoints to log daily expenses, categorize spending, and generate basic financial summaries.",
        "title": "API Backend Expense",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/transfers": {
            "get": {
                "description": "get all transfer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transfers"
                ],
                "summary": "List transfer",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response-any"
                        }
                    }
                }
            },
            "post": {
                "description": "create transfer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transfers"
                ],
                "summary": "Add new transfer",
                "parameters": [
                    {
                        "description": "Create new transfer payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateTransferPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response-any"
                        }
                    }
                }
            }
        },
        "/transfers/{id}": {
            "get": {
                "description": "get one filters by transfer id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transfers"
                ],
                "summary": "Get an transfer",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Transfer ID (Optional)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response-any"
                        }
                    }
                }
            },
            "put": {
                "description": "update transfer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transfers"
                ],
                "summary": "Edit transfer",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Transfer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Edit transfer payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateTransferPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response-any"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete transfer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transfers"
                ],
                "summary": "Remove transfer",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Transfer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response-any"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "apperror.AppError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.CreateTransferPayload": {
            "type": "object",
            "required": [
                "amount",
                "destination_account",
                "from_account",
                "title",
                "transfer_date"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "destination_account": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                },
                "from_account": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                },
                "note": {
                    "type": "string",
                    "maxLength": 255
                },
                "title": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                },
                "transfer_date": {
                    "type": "string",
                    "example": "2006-01-02"
                }
            }
        },
        "dto.Meta": {
            "type": "object",
            "properties": {
                "paginate": {
                    "$ref": "#/definitions/dto.Paginate"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "dto.Paginate": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "next": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "prev": {
                    "type": "string"
                },
                "total_pages": {
                    "type": "integer"
                },
                "total_rows": {
                    "type": "integer"
                }
            }
        },
        "dto.Response-any": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/apperror.AppError"
                },
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/dto.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "dto.UpdateTransferPayload": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "destination_account": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                },
                "from_account": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                },
                "note": {
                    "type": "string",
                    "maxLength": 255
                },
                "title": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                },
                "transfer_date": {
                    "type": "string",
                    "example": "2006-01-02"
                }
            }
        }
    }
}